requirements.txt
passlib[bcrypt] 
python-jose[cryptography] 
python-dotenv
fastapi
pydantic[email]
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
databases
python-dotenv
psycopg2-binary 
supabase
bcrypt==4.0.1
